Separate navbar component and show active page
